Add files via upload
main objective of this C++ project on Mess Management System is to manage the items available in the mess and their prices per plate.  
It manages all the information about items, orders and their prices.  
The project is built for the billing purpose at the customer level as it reduces the time for the billing management and the manual work for the managing items and orders. 
It tracks all the details about the Mess The, food available and their prices as well as the bill generation.